Set the SQL Float precision
By default mysql use a precision of 23 to store a Float and
sqlite/postgresql use 53.

The patch set the precision of Float to 53 to force all sql backends to
use the same double precision.

+++ ceilometer/storage/sqlalchemy/models.py
@@ -186,7 +186,7 @@ class MetaFloat(Base):
     )
     id = Column(Integer, ForeignKey('meter.id'), primary_key=True)
     meta_key = Column(String(255), primary_key=True)
-    value = Column(Float, default=False)
+    value = Column(Float(53), default=False)
 
 
 class Meter(Base):
@@ -377,7 +377,7 @@ class Trait(Base):
     trait_type = relationship("TraitType", backref=backref('trait_type'))
 
     t_string = Column(String(255), nullable=True, default=None)
-    t_float = Column(Float, nullable=True, default=None)
+    t_float = Column(Float(53), nullable=True, default=None)
     t_int = Column(Integer, nullable=True, default=None)
     t_datetime = Column(PreciseTimestamp(), nullable=True, default=None)
